Compress pose estimation data modules to one file
