--- src/netspresso_trainer/dataloaders/pose_estimation/__init__.py
@@ -1,2 +1,0 @@
-from .dataset import PoseEstimationSampleLoader
-from .local import PoseEstimationCustomDataset
